Create Activity6.c
What's up! That's my first time using if structure. The original activity only wanted the mean between the two grades, but since I already did that in my fourth or fifth activity, I decided to make this one here harder. It shows also if the student got an A, B, C...